fix: consolidate runtime host files to satisfy pkg-file-count limit
Merge child host primitive implementations and fixture-backed tests into
host_children.go and host_children_test.go, and combine pipeline/policy
tests into host_pipeline_policy_test.go so the runtime package stays within
the 15-file pkg-file-count gate required by Build, Lint, and API CI.
